--- app/src/main/java/com/ufscar/mobile/hestiaapp/CadastraImovelActivity.kt
@@ -1,76 +1,0 @@
-package com.ufscar.mobile.hestiaapp
-
-import android.content.Intent
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.text.TextUtils
-import com.google.firebase.auth.FirebaseAuth
-import com.ufscar.mobile.hestiaapp.util.FirestoreImovelUtil
-import kotlinx.android.synthetic.main.fragment_cadastra_info_imovel.*
-import org.jetbrains.anko.clearTask
-import org.jetbrains.anko.newTask
-
-class CadastraImovelActivity : AppCompatActivity() {
-
-    private val REQUIRED_FIELD = "Campo obrigatório"
-
-//    lateinit var partOneFragment: CadastraInfoImovelFragment
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_cadastra_imovel)
-
-//        prosseguir.setOnClickListener {
-//            val cadastraFotos = CadastroFotosImovelFragment()
-//            val transaction = supportFragmentManager.beginTransaction()
-//            transaction.replace(R.id.fragment_container, cadastraFotos)
-//            transaction.addToBackStack(null)
-//            transaction.commit()
-//        }
-
-        prosseguir.setOnClickListener {
-           //FIXME: o erro só aparece no último campo do teste
-            if(verifyRequiredFields()) {
-                FirestoreImovelUtil.insertImovel(
-                        edtTipo.text.toString(),
-                        edtMoradores.text.toString().toInt(),
-                        0,
-                        0,
-                        edtPreco.text.toString().toInt(),
-                        edtQuartos.text.toString().toInt(),
-                        edtBanheiros.text.toString().toInt(),
-                        edtSalas.text.toString().toInt(),
-                        edtVagas.text.toString().toInt(),
-                        edtDesc.text.toString(),
-                        edtEnd.text.toString(),
-                        FirebaseAuth.getInstance().uid ?: "")
-
-                val intentDono = Intent(this, MeusImoveisActivity::class.java)
-                intentDono.newTask().clearTask()
-                    startActivity(intentDono)
-                    finish()
-            }
-        }
-    }
-
-    fun verifyRequiredFields(): Boolean {
-        var result = true
-        if(TextUtils.isEmpty(edtEnd.text)) {
-            edtEnd.error = REQUIRED_FIELD
-            result = false
-        }
-        if (TextUtils.isEmpty(edtPreco.text)) {
-            edtPreco.error = REQUIRED_FIELD
-            result = false
-        }
-        if (TextUtils.isEmpty(edtTipo.text)) {
-            edtTipo.error = REQUIRED_FIELD
-            result = false
-        }
-        if (TextUtils.isEmpty(edtDesc.text)) {
-            edtDesc.error = REQUIRED_FIELD
-            result = false
-        }
-        return result
-    }
-}
