+++ app/src/main/java/com/ufscar/mobile/hestiaapp/cenarios/info_imovel/InfoImovelContract.kt
@@ -1,0 +1,14 @@
+package com.ufscar.mobile.hestiaapp.cenarios.info_imovel
+
+import android.content.Context
+
+interface InfoImovelContract {
+    interface View {
+        fun showFabEdit()
+        fun showFabFavorite()
+    }
+
+    interface Presenter {
+        fun onLoadFab(context: Context)
+    }
+}
