+++ app/src/main/java/com/ufscar/mobile/hestiaapp/cenarios/meus_imoveis/MeusImoveisContract.kt
@@ -1,0 +1,14 @@
+package com.ufscar.mobile.hestiaapp.cenarios.meus_imoveis
+
+import com.ufscar.mobile.hestiaapp.model.Imovel
+
+interface MeusImoveisContract {
+    interface View {
+        fun showList(list: ArrayList<Imovel>)
+        fun hideLoading()
+    }
+
+    interface Presenter {
+        fun onListUpdate()
+    }
+}
