+++ app/src/main/java/com/ufscar/mobile/hestiaapp/util/GlideModule.kt
@@ -1,0 +1,18 @@
+package com.ufscar.mobile.hestiaapp.util
+
+import android.content.Context
+import com.bumptech.glide.Glide
+import com.bumptech.glide.Registry
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+import com.firebase.ui.storage.images.FirebaseImageLoader
+import com.google.firebase.storage.StorageReference
+import java.io.InputStream
+
+@GlideModule
+class GlideModule : AppGlideModule() {
+    override fun registerComponents(context: Context, glide: Glide, registry: Registry) {
+        registry.append(StorageReference::class.java, InputStream::class.java,
+                FirebaseImageLoader.Factory())
+    }
+}
