+++ app/src/main/res/layout/activity_info_imovel.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".InfoImovelActivity">
+    tools:context=".cenarios.info_imovel.InfoImovelActivity">
 
     <ScrollView
         android:layout_width="match_parent"
