+++ app/src/main/res/layout/activity_perfil.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PerfilActivity">
+    tools:context=".cenarios.perfil.PerfilActivity">
 
     <ProgressBar
         android:id="@+id/progress_bar"
